put tests back in the right place after merge
